[BUGFIX] Fix autoloading, make ext work again
Fix autoloading so that the perfectlightbox class can be found again.

Also fixes the problem that if click-enlarge is activated, the word
"Array" is printed in front of each image, the image is not linked and
no lightbox appears.

This makes the extension compatible with current versions of TYPO3
again.

Fixes #1
Fixes #2
Fixes #3